chore(blockifier_reexecution): move serde functions to dedicated file
