+++ crates/blockifier_reexecution/src/state_reader.rs
@@ -5,5 +5,6 @@ pub mod raw_rpc_json_test;
 #[cfg(test)]
 #[cfg(feature = "blockifier_regression_https_testing")]
 pub mod rpc_https_test;
+pub mod serde_utils;
 pub mod test_state_reader;
 pub mod utils;
